Ora funziona anche per i numeri di tipo "long".

+++ EquazioniLibrary/EquazioniTest.cs
@@ -8,15 +8,17 @@ namespace EquazioniLibrary
 {
     public static class EquazioniTest
     {
-        public static int Fattoriale(int num)
+        public static long Fattoriale(int num)
         { 
-            int fattoriale=1;
+            long fattoriale=1;
             
             for(int i = num;i>0;i--)
             {
               fattoriale = fattoriale * i;
             }
+            
             return fattoriale;
         }
     }
 }
+
